requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
facenet-pytorch==2.2.9
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.1.2
requests==2.24.0
six==1.15.0
torch==1.5.1
torchvision==0.6.1
urllib3==1.25.9
Werkzeug==1.0.1
wincertstore==0.2
Flask-API==2.0
